moved bitmap code to cc file. added cast to unsigned type

+++ source/HQ2x.cc
@@ -72,7 +72,7 @@ uint32_t *HQ2x::resize(
 
 		// adjusts the previous and next line pointers
 		if (row > 0)
-			previous = -width;
+			previous = -(int)width;
 		else
 		{
 			if (wrapY)
@@ -85,7 +85,7 @@ uint32_t *HQ2x::resize(
 		else
 		{
 			if (wrapY)
-				next = -(width * (height - 1));
+				next = -(int)(width * (height - 1));
 			else
 				next = 0;
 		}
